refactor(mcp): extract blocking tool dispatch
