+++ crates/rag-rat-mcp/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod agent_hook;
+mod blocking;
 pub mod server;
 pub mod tools;
 #[cfg(unix)]
